pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
